pyproject.toml
[project]
name            = "duckdb-hybrid-doc-search"
version         = "0.1.0"
description     = "Hybrid FTS + VSS search over Markdown using DuckDB"
requires-python = ">=3.10"
license         = { text = "MIT" }
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "banks==2.1.2",
    "black==25.1.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "coverage==7.8.0",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "dirtyjson==1.0.8",
    "duckdb==1.2.2",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "greenlet==3.2.1",
    "griffe==1.7.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "lindera-py>=0.42.0",
    "llama-index-core==0.12.33.post1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mcp==1.6.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mypy==1.15.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "orjson==3.10.16",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "propcache==0.3.1",
    "protobuf==6.30.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pytest==8.3.5",
    "python-dotenv==1.1.0",
    "python-ulid==3.0.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "ruff==0.11.7",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==4.1.0",
    "sentencepiece==0.2.0",
    "setuptools==80.0.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "sse-starlette==2.3.3",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "torch==2.7.0",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "wrapt==1.17.2",
    "yarl==1.20.0",
]

[project.scripts]
duckdb-hybrid-doc-search = "duckdb_hybrid_document_search.cli:app"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]

[tool.mypy]
strict = true
python_version = "3.10"
